refactor: split tracks, genres api modules
- enable `react-refresh/only-export-components` rule & resolve appeared warnings for shadcn button and badge
- `TracksManager` -> `MusicManager`, move it to a separate file in the component folder at the highest level
