+++ src/components/tracks/dialogs/base-form/index.tsx
@@ -4,8 +4,8 @@ import { FormField } from "@/components/tracks/dialogs/base-form/field";
 import { GenreSelector } from "@/components/tracks/dialogs/base-form/genres-selector";
 import { Button } from "@/components/ui/button";
 import { Label } from "@/components/ui/label";
+import { type Genres } from "@/lib/api/genres";
 import { type CreateTrackFormData } from "@/lib/api/tracks";
-import { type Genres } from "@/lib/api/types";
 
 interface BaseFormProps {
   form: UseFormReturn<CreateTrackFormData>;
