+++ src/lib/api/client.ts
@@ -1,4 +1,13 @@
-import { createApiClient } from "../../generated/api";
+import { type ApiOf } from "@zodios/core";
+
+import {
+  createApiClient,
+  schemas as generatedSchemas,
+  type api,
+} from "@/generated/api";
+
+export type Api = ApiOf<typeof api>;
+export const apiSchemas = generatedSchemas;
 
 export const API_BASE_URL =
   import.meta.env.VITE_API_URL || "http://localhost:8000";
