+++ src/lib/api/genres/index.ts
@@ -1,0 +1,3 @@
+export * from "./hooks";
+export * from "./service";
+export * from "./types";
