+++ src/lib/api/genres/types.ts
@@ -1,0 +1,6 @@
+import { type ZodiosResponseByAlias } from "@zodios/core";
+
+import { type Api } from "../client";
+
+export type Genres = ZodiosResponseByAlias<Api, "getApigenres">;
+export type Genre = Genres[number];
